third_party/blink/web_tests/html/sections/article-element-expected.txt

Various tests for the article element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<article> closes <p>:
PASS article1.parentNode.nodeName == "p" is false
<p> does not close <article>:
PASS p1.parentNode.nodeName is "ARTICLE"
<article> can be nested inside <article>:
PASS article3.parentNode.id is "article2"
Residual style:
PASS getWeight("article4") is "700"
PASS getWeight("span1") is "700"
FormatBlock:
PASS document.getElementById("span2").parentNode.nodeName is "ARTICLE"
PASS successfullyParsed is true

TEST COMPLETE
